feat(gulps): Moved readTextFile, readJsonFile, filterFiles from release to helper
